fix out again and commands
