+++ input.txt
@@ -1,4 +1,4 @@
  test first
- sencond str
+ 6 sencond 1 2 3 str 4 5
  third str nice
  1 2 3 4 5 6 7 8 9 10 10 10 1
